## [v8.1.0]() (2018-06-10)
**Added**
- added `blog` data type. Various methods in `EventRegistry` class accept it, such as `suggestNewsSources()`, `suggestSourcesAtPlace()` and `getNewsSourceUri()`.

**Updated**
- `QueryArticlesIter.initWithComplexQuery()` now accepts also the `dataType` parameter (by default `news`).

**Removed**
- Removed the parameter `articleBatchSize` from `QueryArticlesIter.execQuery` since it was not useful. We are always returning the maximum number of results that can be obtained with a single query.

+++ eventregistry/tests/TestQueryArticles.py
@@ -30,7 +30,7 @@ class TestQueryArticles(DataValidator):
         expectedCount = iter.count(self.er)
 
         countPerPage = 20000
-        pages = math.ceil(expectedCount / countPerPage)
+        pages = int(math.ceil(expectedCount / countPerPage))
         q = QueryArticles(conceptUri=self.er.getConceptUri("germany"))
         items = []
         for page in range(1, pages+1):
